requirements.txt
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
dnspython==2.6.1
exceptiongroup==1.2.0
fastapi==0.110.0
h11==0.14.0
idna==3.6
importlib_metadata==7.1.0
macholib==1.16.3
motor==3.3.2
packaging==24.0
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
PyJWT==2.8.0
pymongo==4.6.2
python-dotenv==1.0.1
python-multipart==0.0.9
redis==5.0.2
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.27.1
websockets==12.0
zipp==3.18.1
